rest_api/requirements.txt
flask-lambda
boto3
requests
aws-psycopg2
flask-cors
smart_open[s3]